+++ docs/changelog.md
@@ -2,7 +2,11 @@
 
 Versions follow [Semantic Versioning](https://semver.org/): `<major>.<minor>.<patch>`.
 
-* Change dependency behaviour when a case or module name does not exist. 
+* Add prohibition to run tests on the port if the port is busy.
+  [[PR-136](https://github.com/everypinio/hardpy/pull/136)]
+* Add marker **critical**.
+  [[PR-135](https://github.com/everypinio/hardpy/pull/135)]
+* Change dependency behaviour when a case or module name does not exist.
   Such a test or module will be executed.
   [[PR-130](https://github.com/everypinio/hardpy/pull/130)]
 * Add the ability to add dependency from multiple tests.